+++ lib/td_qx/data_views/queryable_properties/select_field.ex
@@ -29,4 +29,14 @@ defmodule TdQx.DataViews.QueryableProperties.SelectField do
     |> cast_embed(:expression, with: &Expression.changeset_for_aggregate/2, required: true)
     |> validate_required([:alias, :id])
   end
+
+  def unfold([%__MODULE__{} | _] = fields) do
+    Enum.map(fields, &unfold/1)
+  end
+
+  def unfold([]), do: []
+
+  def unfold(%__MODULE__{alias: field_alias, expression: expression}) do
+    %{__type__: "select_field", alias: field_alias, expression: Expression.unfold(expression)}
+  end
 end